+++ src/cpprest-client/model/Subsystem.h
@@ -56,20 +56,20 @@ public:
     /// <summary>
     /// Unix timestamp when this entity was created.
     /// </summary>
-    int64_t getCreatedAt() const;
+    int32_t getCreatedAt() const;
     bool createdAtIsSet() const;
     void unsetCreatedAt();
 
-    void setCreatedAt(int64_t value);
+    void setCreatedAt(int32_t value);
 
     /// <summary>
     /// The unique identifier of this entity.
     /// </summary>
-    int64_t getId() const;
+    int32_t getId() const;
     bool idIsSet() const;
     void unsetId();
 
-    void setId(int64_t value);
+    void setId(int32_t value);
 
     /// <summary>
     /// The label value of the subsystem.
@@ -83,21 +83,21 @@ public:
     /// <summary>
     /// Unix timestamp when this entity was last updated.
     /// </summary>
-    int64_t getUpdatedAt() const;
+    int32_t getUpdatedAt() const;
     bool updatedAtIsSet() const;
     void unsetUpdatedAt();
 
-    void setUpdatedAt(int64_t value);
+    void setUpdatedAt(int32_t value);
 
 
 protected:
-    int64_t m_CreatedAt;
+    int32_t m_CreatedAt;
     bool m_CreatedAtIsSet;
-    int64_t m_Id;
+    int32_t m_Id;
     bool m_IdIsSet;
     utility::string_t m_Name;
     bool m_NameIsSet;
-    int64_t m_UpdatedAt;
+    int32_t m_UpdatedAt;
     bool m_UpdatedAtIsSet;
 };
